backend/requirements.txt
fastapi>=0.110
uvicorn>=0.30
pydantic[email]>=2.0
python-jose[cryptography]>=3.3
passlib[bcrypt]>=1.7
supabase @ file:./vendor/supabase_stub
redis>=5.0
arq>=0.26
python-dotenv>=1.0
httpx>=0.27
pydantic-settings>=2.0
docker>=7.0.0

